refactor(Rank): 상금 타입을 long으로 수정
- Rank : 상금 타입을 long으로 수정
- GameResult : 변경된 상금 타입에 맞게 수정
- GameResultTest : 변경된 GameResult에 맞게 수정

+++ src/main/java/lotto/domain/GameResult.java
@@ -10,11 +10,11 @@ public class GameResult {
         this.statistics = new EnumMap<>(statistics);
     }
 
-    public int getTotalPrize() {
-        int totalPrize = 0;
+    public long getTotalPrize() {
+        long totalPrize = 0L;
 
         for (Rank rank : Rank.values()) {
-            int rankPrize = rank.getPrize() * getCount(rank);
+            long rankPrize = rank.getPrize() * getCount(rank);
             totalPrize += rankPrize;
         }
